requirements.txt
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
decorator==5.1.1
Django==4.2.10
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
executing==2.0.1
idna==3.6
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
jwcrypto==1.5.3
matplotlib-inline==0.1.6
mysqlclient==2.2.4
oauthlib==3.2.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyMySQL==1.1.0
pytz==2024.1
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
